fix: rename number namespace to num
