+++ src/utils/hash.ts
@@ -9,7 +9,7 @@ import { felt } from './calldata/cairo';
 import { starkCurve } from './ec';
 import { addHexPrefix, buf2hex, removeHexPrefix, utf8ToArray } from './encode';
 import { parse, stringify } from './json';
-import { BigNumberish, isHex, isStringWholeNumber, toBigInt, toHex, toHexString } from './number';
+import { BigNumberish, isHex, isStringWholeNumber, toBigInt, toHex, toHexString } from './num';
 import { encodeShortString } from './shortString';
 
 export const transactionVersion = 1n;
