+++ src/utils/responseParser/rpc.ts
@@ -9,7 +9,7 @@ import {
   GetTransactionResponse,
   RPC,
 } from '../../types';
-import { toBigInt } from '../number';
+import { toBigInt } from '../num';
 import { ResponseParser } from '.';
 
 type RpcGetBlockResponse = RPC.GetBlockWithTxHashesResponse & {
